Dynamic properties (#564)
This PR adds the ability to dynamically change properties at runtime.
The key parts are targeting content(s) with a KeyPath and then updating their value either with a LottieStaticValueCallback for a single fire and forget value or a LottieValueCallback which will be called with the current animation value information every time the animation is rendered. See the docs for relevant classes for more info.