Ignore null parent layers
It isn't clear what the actual issue is here and it may actually
be a bug in d8. This should at least prevent the crash.
Worst case, in these rare cases, a layer may not be parented correctly
but it could be resolved in future d8 updates as well.
Fixes #524