Enforce order of operations to avoid rounding errors (#1897)
Previously, the calculation of endLength could suffer from rounding errors. While most of the time, these errors are not a big deal, when the end value is 100, this can result in endLength being larger than totalLength, causing unexpected behavior where the animation reverts to an initial state when it should be showing the final state. By forcing the order of operations, we can get the answer we expect / need.

Co-authored-by: Scott Weber <sweber@duosecurity.com>