Move snapshot tests into their own app and improve performance (#1914)
The dedicated module prevents extra code in the main sample app modules and allows for easy creation of compose snapshot tests. I also found a bug in which the rendering of animations was not fan-in when it should have been. This increased the performance of the test suite locally by ~4x.